new basic plugin functionality
when a post becomes published, increment all of the term post counts
for associated terms

when a post becomes un-published, deferment the same, never dipping
below zero

add the LTCU_post_quick_update_terms_count action at the end of the
operation.  This can be used (for example) to detect bad behavior,
including posts with very large numbers of terms.