Adding edit(ed)_term_taxonomy actions and a unit test for testing that they are being fired for all terms and taxonomies
Since we are updating more tt_ids at once, we need to fire the actions inside a loop. Not the best way to do it, but it seems to be necessary.

The test for testing the action is inspired by https://core.trac.wordpress.org/browser/trunk/tests/phpunit/tests/actions/closures.php when it comes to using the $GLOBALS