Added if statement to take care of an edge condition in LTCU_Plugin->quick_update_terms_count
Problem: If the terms passed to the function LTCU_Plugin->quick_update_terms_count is already processed it gives an empty array to the cache purging function which processes all the terms. This is both time intesive and not the required behaviour.

Fix: Added code to return the funtion when the terms are already processed